[TIDL TDA2x-TDA3x] Added linux compiled libraries for TI-DL
Linux built libraries are not included by default in TI-DL software release package.
Added only the X86 version since the x86-64 is not functional.
Note that if one wants to build them from the TI-DL
release package or from the dsp_apps repo, one needs to edit the dsp_apps/makerules/rules.mk
file and set CC=gcc -c instead of CC=g++ -c otherwise the generated library
won't be compatible with the rest of ti-ovx library.
Changing the compiler from g++ to gcc will result in compiler error of TI-DL so
it's better to cherry-pick the commit from the dsp_apps repos that fix these issues before building.
Most of the errors are simple to fix, such as replacing 'inline' with 'static inline' and
also replacing '#include <cfloat>' with '#include <float.h>'.