[TI-DL DSP] Workaround for missing cache write back in TI-DL library
There is a missing cache-write back call in the TIDL_ativate() function of TI-DL.
This can call issue when running on the target.
This workaround offsets this issue by adding a call to tivxMemBufferUmap() after the TIDL_init()
function so a call to cache-write back in TIDL_activate() is no longer required.