[TIOVX-596] On TDA2x/TDA3x bios build, allocation in TIVX_MEM_INTERNAL_L3 heap fails for DSP
The TIVX_MEM_INTERNAL_L3 heap was mapped to UTILS_HEAPID_OCMC_SR
which doesn't work for DSP on TDA2x/3x .
The fix is to map it to UTILS_HEAPID_DDR_CACHED_SR .

This affects only algvision kernels that makes use of
TIVX_MEM_INTERNAL_L3 heap. So far only TI-DL on DSP is affected.