[TIOVX-598] vx_tensor object shall have new attributes: scaling_divisor and scaling_divisor_fixed_point_position
These attributes are useful in the context of TI-DL.
In TI-DL, there is a need to pass a parameter dataQ for each input and output
buffer. This dataQ parameter is a divisor scaling factor that must be applied
to each element of the input or output tensors in order to recover the real
values. If this dataQ becomes part of the tensor object itself,
it will be implicitly passed along the tensor and there is not need to
explicitly pass the parameter dataQ to tivxTIDLNode(), simplifying its
interface and the application code using it.