[TIOVX-599] alg_vision memory allocation needs to be optimized for records that have scratch attribute
Currently scratch memory records are allocated as persistent using tivxMemAlloc(). This is not optimum in case multiple ivision nodes need memory records in the same heap, such as DSP L2 memory. Attribute scratch should allow allocating overlapping memories but tivxMemAlloc() will give
consecutive memories, leading to out of memory issue very quickly.

This issue was observed in the context of TI-DL object detection network
in which there are 4 graphs,  each one containing a EVE node and a DSP node
calling TI-DL. The issue was not observed with EVE nodes, because we use EVE1,
 EVE2, EVE3, EVE4. But for DSP, all the nodes use DSP1.
The memory allocation in L2 would fail for the DSP1 node in the second graph.

In case of L2, a simple solution would be to free the memory right after
being allocated. In case of external, a memory management not relying on
tivxMemAlloc() should be implemented.

For now, we just need to worry about L2 as there is plenty of external memory
available.